+++ tests/test_helpers.py
@@ -55,11 +55,6 @@ def test_get_convert_dist():
     assert f(1) == 1.609_344
 
 
-def test_get_utm_crs():
-    assert gkh.get_utm_crs(lat=10, lon=-175) == "EPSG:32601"
-    assert gkh.get_utm_crs(lat=-36.840556, lon=174.4) == "EPSG:32760"
-
-
 def test_get_segment_length():
     s = sg.LineString([(0, 0), (1, 0)])
     p = sg.Point((1 / 2, 0))